show time

+++ fuzz_red_black_tree.c
@@ -244,8 +244,8 @@ int main() {
   start = clock();
   current = clock();
   i = 0;
-  while (((double)(current-start) / CLOCKS_PER_SEC) < TIMEOUT) {
-    printf("TIME: %d", (int)((double)(current-start)/CLOCKS_PER_SEC));
+  while (((double)(current-start) / (double)CLOCKS_PER_SEC) < TIMEOUT) {
+    printf("TIME: %d", (int)((double)(current-start)/(double)CLOCKS_PER_SEC));
     current = clock();
     fuzzit();
     i++;
